Use a Swift-specific build script
