+++ package.json
@@ -70,6 +70,7 @@
     "palette:build:css:variables:hex": "node builders/scss-variables-from-json > dist/color-variables.scss",
     "palette:build:css:variables:rgb": "node builders/scss-variables-rgb-from-json > dist/color-variables-rgb.scss",
     "palette:build:ios": "node builders/ios-assets-from-json",
+    "palette:build:swift": "node builders/swift-raw-data-from-json",
     "palette:build:json": "yarn palette:build:json:flat && yarn palette:build:json:meta && yarn palette:build:json:meta-debug",
     "palette:build:json:flat": "node builders/json-flat-from-json > dist/colors.json",
     "palette:build:json:meta": "node builders/json-meta-from-json > dist/colors.meta.json",
